Refactors API
