+++ src/lib.rs
@@ -2,20 +2,22 @@
 // src/lib.rs
 uniffi::setup_scaffolding!();
 
-mod api;
 mod bozorth;
 mod consts;
 mod encoding;
 mod errors;
+mod extractor;
 pub(crate) mod ffi;
 mod imutils;
 mod minutia;
 mod minutiae;
+mod sivv;
+mod structs;
 
-pub use api::{
-    extract_minutiae, extract_minutiae_from_image_file, load_iso_19794_2_2005, NfiqQuality, Point,
-    ROI,
-};
+pub use structs::{NbisExtractorSettings, NfiqQuality, Point, ROI};
+
 pub use errors::NbisError;
+pub use extractor::new_nbis_extractor;
+pub use extractor::NbisExtractor;
 pub use minutia::{Minutia, MinutiaKind, Position};
 pub use minutiae::Minutiae;
