+++ src/structs/mod.rs
@@ -1,0 +1,13 @@
+mod extractor_settings;
+mod nfiq_quality;
+mod nfiq_result;
+mod point;
+mod roi;
+mod sivv_result;
+
+pub use extractor_settings::NbisExtractorSettings;
+pub use nfiq_quality::NfiqQuality;
+pub(crate) use nfiq_result::NfiqResult;
+pub use point::Point;
+pub use roi::ROI;
+pub(crate) use sivv_result::SIVVResult;
